requirements.txt
django-admin
requests
SQLAlchemy
beautifulsoup4
pandas 